+++ test/.eslintrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['carpages/rules/dsp', 'plugin:qunit/recommended', 'plugin:qunit/two'],
+  plugins: ['qunit']
+};